refactor master, proposer, etc...
